Fix compilation warning due to different types in min()
In file included from emlog.c:33:0:
emlog.c: In function ‘read_from_emlog’:
include/linux/kernel.h:672:17: warning: comparison of distinct pointer types lacks a cast
  (void) (&_min1 == &_min2);  \
                 ^
emlog.c:238:15: note: in expansion of macro ‘min’
     *length = min((loff_t)*length, EMLOG_FIRST_EMPTY_BYTE(einfo) - *offset);
               ^

The warning was introduced in 560e37bef89181c20d5214102f41cb8698236d3e.

Reported-by: Simon Boulet <simon@nostalgeek.com>

+++ emlog.h
@@ -33,9 +33,9 @@ struct emlog_info {
     char *data;                 /* The circular buffer data */
     size_t size;                /* Size of the buffer pointed to by 'data' */
     int refcount;               /* Files that have this buffer open */
-    int read_point;             /* Offset in circ. buffer of oldest data */
-    int write_point;            /* Offset in circ. buffer of newest data */
-    int offset;                 /* Byte number of read_point in the stream */
+    size_t read_point;          /* Offset in circ. buffer of oldest data */
+    size_t write_point;         /* Offset in circ. buffer of newest data */
+    loff_t offset;              /* Byte number of read_point in the stream */
     struct emlog_info *next;
 };
